fix unexport issue

+++ src/main/java/beer/gpio/device/BasePin.java
@@ -1,5 +1,6 @@
 package beer.gpio.device;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 
@@ -14,11 +15,17 @@ public abstract class BasePin {
 	public BasePin(int pinNumber, Direction pinDirection) throws TemperatureSensorException {
 		this.pinNumber = pinNumber;
 		
-		writeFile(getFilePaths().getUnexportPath(), pinNumber + "");
+		if (alreadyExported()) {
+			writeFile(getFilePaths().getUnexportPath(), pinNumber + "");
+		}
 		writeFile(getFilePaths().getExportPath(), pinNumber + "");
 		writeFile(getFilePaths().getDirectionPath(pinNumber), pinDirection.getValue());
 	}
 	
+	private boolean alreadyExported() {
+		return new File(getFilePaths().getDirectionPath(pinNumber)).exists();
+	}
+
 	public FilePaths getFilePaths() {
 		return filePaths;
 	}
@@ -32,6 +39,7 @@ public abstract class BasePin {
 	}
 
 	protected void writeFile(String fileName, String value) throws TemperatureSensorException {
+		System.out.println("writeFile: " + fileName + ", " + value); 
 		try {
 			// TODO: doesn't throw an exception if the file is not there!
 			FileOutputStream fos = new FileOutputStream(fileName);
